rename routes to filters for consistency with warp
