+++ src/filters/geoprocessing/get_linestring.rs
@@ -1,6 +1,6 @@
 
 use crate::data::IndexedData;
-use crate::routes::query_parameters::{QueryParametersLine, output_format::OutputFormatLines};
+use crate::filters::query_parameters::{QueryParametersLine, output_format::OutputFormatLines};
 use nickslinetoolsrust::line_string_measured::{LineStringMeasured};
 use crate::helpers::{convert_metres_to_degrees, ErrorWithStaticMessage};
