+++ src/filters/mod.rs
@@ -19,3 +19,6 @@ pub use custom_rejection_handler::custom_rejection_handler;
 
 mod load_data_and_combine_filters;
 pub use load_data_and_combine_filters::load_data_and_combine_filters;
+
+mod with_shared_data;
+pub use with_shared_data::with_shared_data;
