+++ src/filters/query_parameters/query_parameters_point.rs
@@ -55,7 +55,6 @@ mod tests {
     #[test]
     /// Test when all parameters are provided
     fn test_query_parameters_point_all() {
-        let query = "road=H001&slk=5&cwy=SL&offset=10&f=wkt";
         let query: QueryParametersPoint = serde_urlencoded::from_str(query).unwrap();
         assert_eq!(query, QueryParametersPoint{
             road: "H001".to_string(),
