+++ src/main.rs
@@ -2,7 +2,7 @@ use std::{net::SocketAddr, error::Error};
 use clap::Parser;
 
 mod helpers;
-mod routes;
+mod filters;
 mod data;
 mod settings;
 
@@ -13,7 +13,7 @@ async fn main() -> Result<(), Box<dyn Error>> {
     // see run with --help to see nice instructions generated by clap
     let settings = settings::Settings::parse();
 
-    let filter = routes::load_data_and_combine_filters(&settings).await?;
+    let filter = filters::load_data_and_combine_filters(&settings).await?;
 
     // Serve
     let address:SocketAddr = (settings.NLR_ADDR, settings.NLR_PORT).into();
